requirements.txt
annotated-types==0.7.0
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysqlclient==2.2.6
psycopg2==2.9.10
pydantic==2.10.1
pydantic_core==2.27.1
PyJWT==2.10.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
